requirements.txt
flask
flask-cors
google-genai
pinecone
cohere
python-dotenv
pypdf
python-docx
gunicorn